// src/inspections/enums/inspection-validity-rule.enum.ts
export enum InspectionValidityRule {
  NEW_FAMILY_VEHICLE = 'NEW_FAMILY_VEHICLE',
  FAMILY_VEHICLE_UP_TO_4_PERSONS_RENEWAL = 'FAMILY_VEHICLE_UP_TO_4_PERSONS_RENEWAL',
  FAMILY_VEHICLE_5_TO_9_PERSONS_NON_COMMERCIAL_RENEWAL = 'FAMILY_VEHICLE_5_TO_9_PERSONS_NON_COMMERCIAL_RENEWAL',
  NEW_COMMERCIAL_PASSENGER_OR_LIGHT_GOODS_VEHICLE = 'NEW_COMMERCIAL_PASSENGER_OR_LIGHT_GOODS_VEHICLE',
  COMMERCIAL_GOODS_UP_TO_ONE_TONNE_RENEWAL = 'COMMERCIAL_GOODS_UP_TO_ONE_TONNE_RENEWAL',
  TRAILER_OR_SEMI_TRAILER = 'TRAILER_OR_SEMI_TRAILER',
  MOTOR_TRICYCLE_OR_MOTORCYCLE_TOWING_TRAILER = 'MOTOR_TRICYCLE_OR_MOTORCYCLE_TOWING_TRAILER',
  TAXI_PASSENGER_5_TO_9_SEATS = 'TAXI_PASSENGER_5_TO_9_SEATS',
  PASSENGER_10_TO_14_SEATS = 'PASSENGER_10_TO_14_SEATS',
  PASSENGER_15_SEATS = 'PASSENGER_15_SEATS',
  PASSENGER_16_TO_20_SEATS = 'PASSENGER_16_TO_20_SEATS',
  PASSENGER_21_OR_MORE_SEATS = 'PASSENGER_21_OR_MORE_SEATS',
  NEW_NON_COMMERCIAL_GOODS_UP_TO_ONE_TONNE = 'NEW_NON_COMMERCIAL_GOODS_UP_TO_ONE_TONNE',
  NON_COMMERCIAL_GOODS_UP_TO_ONE_TONNE_RENEWAL = 'NON_COMMERCIAL_GOODS_UP_TO_ONE_TONNE_RENEWAL',
  NEW_COMMERCIAL_HEAVY_GOODS_OR_PASSENGER = 'NEW_COMMERCIAL_HEAVY_GOODS_OR_PASSENGER',
  GOODS_UP_TO_TWO_TONNES = 'GOODS_UP_TO_TWO_TONNES',
  GOODS_OVER_TWO_TO_FIVE_TONNES = 'GOODS_OVER_TWO_TO_FIVE_TONNES',
  TRACTOR_OR_GOODS_OVER_FIVE_TO_TEN_TONNES = 'TRACTOR_OR_GOODS_OVER_FIVE_TO_TEN_TONNES',
  TRACTOR_OR_GOODS_OVER_TEN_TONNES = 'TRACTOR_OR_GOODS_OVER_TEN_TONNES',
}